Move most sources to a kaleidoscope/ subdir and namespace
Lots of things were already under the `kaleidoscope` namespace, but were in the
"wrong" directory. This big patchset moves everything under `kaleidoscope/`, and
namespaces `Layers_`. There are a few un-namespaced things, but because their
use is so widespread, and they're used so commonly, we don't want to namespace
those (like `handleKeyswitchEvent`), at least not yet.

As a minor side-effect, this moves a few global statics that were once in
`Layers.cpp` into the `Layer_` class itself, for clarity and encapsulation.

Non-namespaced headers that were used outside of Kaleidoscope itself will still
work, but will emit a compile-time warning.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>