LEDs: Guard a few places with #if LED_COUNT > 0
For keyboards without LEDs, `LED_COUNT` is set to zero. Some code assumed it to
be non-zero, which worked fine while they were separate plugins (they weren't
included, so no compilation happened), but now that we're a monorepo, these get
compiled anyway, and emit warnings for these keyboards.

Guard these places with `#if LED_COUNT > 0 ... #endif` to silence the warnings.
The rest of the code isn't guarded in any way, because we still expect people
not to include LED plugins non LED-less keyboards.

This may not be the best solution, but it is one of the most straightforward ones.

Fixes #385.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>