hardware: Introduce KeyboardHardware.has_leds
Intended to make it easier for LED-using code to bail out early, this simple
constant will tell us whether the board has LEDs or if it does not.

Because this will be used regularly, by code not close to the hardware,
`Kaleidoscope.has_leds` is an alias to `KeyboardHardware.has_leds`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>