OneShot: Fix EscapeOneShot so that it cancels stickies too
EscapeOneShot needs to stop bailing out early when a oneshot is pressed (sticky
appears pressed), and continue with its process if there's a sticky oneshot.
This makes stickies cancelable via the EscapeOneShot plugin.

For this to work, OneShot needed an `.isSticky()` method, much like
`.isActive()` and `.isPressed()`.

Fixes #413 and likely addresses part of #408 too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>