Remove now-obsolete idle checks from a few plugins
A long-long time ago, Kaleidoscope used to call `onKeyswitchEvent` (or rather,
it's then-current incarnation) even when keys were idle. For a number of
plugins, this was a case they didn't want to handle, and bailed out early.

When we changed Kaleidoscope to not call the event handlers for idle keys, these
plugins weren't modified at the same time, and the superfluous check remained,
making the firmware both larger and marginally slower.

Since the event handlers are never called when idle, it is safe to remove these
checks.

Affected plugins are `Leader`, `OneShot`, and `TapDance`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>