hardware::Model01: Handle LED index underflow
Initially, we were using `uint8_t` as the type for led indexes, which could not
go below zero, so we only needed one half of a bounds check. Since we switched
to a signed int, we need to check the other half of the boundaries: if the index
is below zero.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>