hardware/Atreus: From select/unselectRow, go to toggleRow
Instead of having a separate function for selecting and unselecting rows, have
only one that toggles them - we only ever want to toggle them anyway. While
there, optimize it slightly, by not toggling `DDRD`: if we set those in `setup`,
they'll remain set, and there's no need to toggle them.

This makes the code cleaner and smaller, at the cost of four bytes of RAM.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>