hardware/ATMegaKeyboard: Only define cRGB if required
If we wish to support hardware that has LEDs, we should not define the `cRGB`
struct and the `CRGB` macro unconditionally. Define them only if the macro is
undefined. This way any hardware that has LEDs, and still wishes to use
`ATMegaKeyboard` as a base can define the macro prior to including the header.

Such hardware will override the LED functions anyway, so there's nothing else we
need to do to support such a scenario.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>