hardware/Atreus: Restore COL_PIN_LIST & key layout to its original
The pin list (and consequently the key layout list) was rearranged in the early
days of the plugin for more optimal scanning code. With -O3 and moving to
`ATMegaKeyboard`, that optimization is no longer relevant. This restores the pin
list and the layout to its original order, the order at which the hardware is
laid out.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>