hardware/ExpanderKeyboard: Base class for boards with I/O expanders
The new base class is based on `ATMegaKeyboard`, and adds support for boards
that use an I/O expander. Slight modifications had to be made to
`ATMegaKeyboard` to allow such subclassing: there's a difference between the
number of matrix rows, and the number of pins used for matrix rows.
ATMegaKeyboard will expose `matrix_rows` as the total number, but will
internally use the pin count to read only a subset of those - leaving the rest
of the expander.

The ExpanderKeyboard class takes advantage of this, and implements the IO
expander on top of it. It is currently based on the ErgoDox code and uses
i2cmaster, but that can be generalized away later, once we have more boards that
use an expander.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>