Deprecate Layer.defaultLayer()
In the original `KeyboardioFirmware`, this was used to store the default keymap
index in EEPROM. When we removed the EEPROM storage, it was meant to be used by
sketches to set a default layer (the minimum layer one can switch to), which was
not nearly as useful. Even worse, the behaviour was complicated to reason about,
and it wasn't used anyway.

For these reasons, it is now deprecated, and will eventually be removed.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>