A new CLI build system
The `kaleidoscope-builder` tool is starting to show its age, and the mistakes
made during its design: it didn't have first-class support for different
hardware, and it was a lot more brittle one would have liked.

The new build system is much more modular, has first-class support for different
hardware, and wraps the tools used much more thinly. It's also less code, too.

This is the Kaleidoscope part of the update, the real deal is in
`Kaleidoscope-Build-Tools`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>