Redial: Make Key_Redial Ranges-based
Add `REDIAL` to `kaleidoscope::ranges`, so the plugin can define `Key_Redial`
itself, and won't need `Redial.key` to be set in the user sketch either. This is
a breaking change, but one that's easy to upgrade to, hence no effort was made
to make it at least partially backwards-compatible.

Fixes #519.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>