Fix the MEH and HYPER aliases
We need to guard the `k` argument, place it within parens, to make sure it is
treated as a single argument, even if it expands to something that would be
ambigous. This is required, because most - if not all - keys are defined as a
list casted to `Key`, which the pre-processor would misinterpret without the
extra parens.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>