LED-Palette-Theme: Flip the component bits upon read & store
EEPROM defaults to `0xff` when uninitialized, which means we'd start off with a
bright white palette. That's a problem, because it draws a lot of power,
especially when all keys are lit bright. For this reason, flip the bits when
reading or storing palette colors, so the default `0xff` becomes `0x00`, black.

Fixes #529.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>