New plugin: UnderglowControl
Inspired by the `LEDControl` plugin, this plugin serves a similar purpose, but
instead of driving per-key RGB LEDs, it is for RGB underglow, as the name
implies.

The API is slightly different than LEDControl's, and is intentionally kept
simple. There is no support for going back and forth between underglow effects
yet, for example.

Hardware that implements underglow, need to set a new define in their main
header: `KALEIDOSCOPE_HARDWARE_HAS_UNDERGLOW`. Without this set, the plugin
becomes a no-op.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>