Repurpose EEPROMSettings' version
We want to be able to notice when the layout of the EEPROM *settings*
changed (which the CRC does not cover). For this reason, we're repurposing the
existing version setting, which wasn't widely used: it is now internal.

We use the version to determine whether the EEPROM has been written to yet, or
if it is uninitialized. This helps us make sure we're starting up with sensible
defaults.

Fixes #559, and fixes #558.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>