OneShot: Remove some obsolete optimizations
In `beforeReportingState()` and `afterEachCycle()` we used to return early if no
OneShots were active. This was easy to do when we were using a bitfield, we'd
just tested for non-zero. Since we're using an array now, this check is more
expensive, and the extra work negates any benefit of returning early.

For this reason, these early returns are removed.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>