Suppressed a compiler warning and added warning suppression macros
This commit silences a warning due to an undefined variable.

A first attemt to locally supress the warning using compiler pragmas was
unsuccessfull due to a compiler bug in all gcc versions. This bug
in all gcc versions below 6.1 makes local diagnostics suppression useless.

As future versions of Arduino will ship with later versions of gcc
from a certain point on, the supplied warning suppression macros will become useful
to avoid the necessity to set global compiler flags like -Wno... and to
enable more precise warning suppression.

The new header file comes with an example that explains how to use
the suppression mechanism.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>