This PR enables logically grouping keys and apply layer switch and toggle operation to selected key groups only. Examples are provide that demonstrate grouping keys by keyboard hand sides, via condition expressions and with keymap-style color arrays.
The basic idea that led to this PR was the question: How to switch one half of the keyboard to another layer while the other half should stay at the base layer. My intention was to switch via the thumb key of the Model01 one half of the keyboard to a modifier layer where all modifiers are supposed to reside on the home row. The other half of the keyboard should stay at the base layer. By means of that I intent to make typing modifier shortcuts with combined modifiers and keys more convenient.

Of coures, this could already be done with the existing Kaleidoscope, e.g. by defining two individual additional layers whose keys for one half match those of the base layer and the other half the modifier layer. The redundancy is obvious.

That's how the idea emerged to modity the layer class in a way that it stores not only one pair of `layer_state_` and `top_active_layer_` but several, each pair of those variables for a separate region of the keyboard. This makes it possible to make separate regions of the keyboard behave as individual keyboards in terms of their layer state. Lacking a better name, I called those keys assigned to such a region a "key group".

Keys can be arbitrarily assigned to key groups. Layer switches/toggles can be defined to affect one or more key groups (see examples).

By default only two key groups are defined. One for the left and one for the right hand side of the keyboard.

* added header `src/Kaleidoscope-KeyGroups.h`
* hardwares implement static constexpr bool isOnLeftHalf(uint8_t row, uint8_t col) convenience methods to test for the hand a key belongs to (one-handed keybords could just unconditionally return true)
* added header `key_groups.h`
* layers class tracks state of up to six key groups now
* all public methods that affect the layer states now can be passed a key group flags to name the affected layers
* method `extern uint8_t groupOfKey(uint8_t row, uint8_t col)` introduced that can be overridden from within the sketch
* fixed Colormap plugin to work with multiple key groups
* fixed plugin LED-ActiveLayerColor to work with multiple key groups
* fixed plugin LED-ActiveModColor to work with multiple key groups
* added TODO-comments in contexts (plugins EEPROM-KeymapOProgrmmer and NumPad) where layer-methods are used but no layer group information is available
* added TODO-comments about the implementation of method `isOnLeftHalf` in some keyboards
* examples were added to illustrate different use cases

The PR will not break any user code or plugins. Only if keygroups are assigned to layer switch operations, the firmware behavior will differ from what it does currently.

Changes will probably not noticeabley affect runtime performance.

Stock firmware: Groth of program size from 25192 to 26992 bytes. Data usage grows from 1414 to 1459 bytes.

Carefully check commit line changes containing TODOs and fix those. Especially concerning plugins EEPROM-KeymapOProgrmmer and NumPad.

This PR comes with quite some changes and both program size and Data usage grow significantly. Nonetheless, this feature might be worth it.

If Arduino eventually would introduce a global per-sketch configuration header (there's an ongoing discussion about this in the Arduino community) we could support several alternative implementations of the layer system, let the user select one and select a default one otherwise.

As this is currently not possible, I chose to "upgrade" the existing layer class.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>