Fix OneShot premature timeout
OneShot keys were failing to activate on the first press because of how the timeout was
implemented. Most of the time, at the end of a cycle, `should_cancel_` was being set to
`true` because the current time was being compared to the last time a OneShot key was
activated, regardless of whether or not a OneShot key was active. As a result, OneShot
keys that were pressed for the first time in a while would fail to register as OneShots.

This change prevents checking the timeout unless there is an active OneShot key.

Fixes #603.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>