Moved LEDMode pointers to PROGMEM
This commit introduces an infrastructure that enables automatic grouping of
plugins of a specific type (e.g. all plugins derived from LEDMode). For
a plugin type group, an array of plugin pointers is automatically generated,
initialzed stored in PROGMEM. Access functions to retreive a typed plugin
by its ID and to query the total amount of plugins of a specific type are
introduced.

Moreover, this commit uses the new infrastructure to eliminate the list
of LEDMode pointers that is maintained by the LEDControl class and stored in
RAM. Explicit registrations of LEDModes with LEDControl is furtheron
no more required. The method LEDControl::mode_add was deprecated
and turned into a noop.

The newly introduced plugin grouping approach was implemented in a generic fashion
that makes it easy to enable the introduction of other plugin type groups.
If e.g. in the future all keyboards are equipped with sound, sensors or whatever
interfaces, where only one plugin at a time is allowed to acquire access,
this is now possible in a straight forward fashion. Just add another
_GENERATE_TYPED_PLUGIN_ARRAY macro invocation to KALEIDOSCOPE_INIT_PLUGINS.

A final remark about specializing plugins:
It is questionable, if it makes sense to add more and more orthogonal
interfaces like LEDMode to derived plugin classes by means of multiple
inheritance, or if it wouldn't be a superior design, to allow derived
plugin classes to have LEDMode members that can be
queried. Pointers to such members could be stored in the same fashion
as it is done by the infrastructure provided by this commit.
This PR, however, does not intend to change this but possibly paves the
way.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>