Switched to linear addressing in most parts of the code
In all places where this is possible I replaced row/col
addressing by directly using linear addressing.

Iterators were added to the MatrixAddr class to enable
range based for loops over all keys or LEDs.

I refrained from implementing a conversion operator
to auto-convert MatrixAddr objects to uint8_t. This is
because of the collisions and thus ambiguous calls with
those LED access functions that work with LED indices.
For those there are already some functions that
accept a uint8_t LED index which is different from
the offset value that is stored in MatrixAddr.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>