Rewrite Qukeys to send only end-of-cycle reports
This change removes the direct dependency Qukeys had on KeyboardioHID (it used
to access the keyboard HID reports directly) by delaying all reports until after
the key scan cycle is finished. There are several corner-case drawbacks to this
approach, but the main one involves the user basically mashing on the keys, so
I'm okay with less-than-perfect output in that case.

The major change is that this will result in Qukeys sending only one report per
cycle (or sometimes two, but only for a single key event), which slows things
down slightly and makes some of the queue handling more complex. However, until
we get a true event-driven framework in Kaleidoscope, this is probably the best
I can do.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>