Add timer helper function
This adds a function to help plugins that use timeouts to efficiently and
correctly check those timeouts by comparing the elapsed time to a timeout
interval without errors due to overflow. It uses the same time value as the
existing `millisAtCycleStart()`, which should be used to set the "start"
timestamp.

Closes #178.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>