Added color names and keygroups
Added colors.h to have the CSS color names in the kaleidoscope::colors namespace.

LEDEffect-SolidColor now has support for using cRGB objects via a new alternate constructor, and works with the color names.

Keygroups.h has been added to offer functions that allow determining if a given key is in a certain group.

Signed-off-by: JD Lien <jd@jdlien.com>