LEDPaletteTheme: Fix the key address refreshAt() uses
In `refreshAt()`, we want to use the key address, instead of the LED address.
`LEDControl` will turn the key address into a LED address itself anyway. This
not only makes the code a tiny bit more efficient, but it also fixes
`refreshAt()`, which was refreshing the wrong key since the conversion to
`KeyAddr`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>