Introducing a new Device API
The new Device API was built on top of composition (instead of inheritance, like
the former one). At the highest level, we have a Device and a DeviceDescription,
the latter has overrides for the various components that make up the device: the
key scanner, LEDs, MCU, and so on.

Many components - like the key scanner and LEDs - also come in a similar setup:
the base class and a description, because this allows us to make them fairly
efficient templates.

With this commit, we lay the foundation of the new API. None of it is being used
yet.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>