driver::bootloader: Implement bootloader drivers
All of our currently supported keyboards have a way to reboot to the bootloader,
to allow themselves to be re-programmed. For a long while, this has been
implemented on a per-keyboard basis, there wasn't a generic solution. The
bootloader drivers aim to fix that situation, by providing a component for
various bootloaders. Devices then can choose which bootloader to use in their
DeviceDescription, and the Device base class will then dispatch to the right
thing.

Fixes #611.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>