device: Let the device select what serial port to use
Instead of delegating the task of deciding whether to use `Serial`, `SerialUSB`
or something else to the FocusSerial plugin, let it stay generic, and add a
`.serialPort()` method to the `Device` class, which will return the appropriate
object when called. By default, the `Device` base class will use a private dummy
object, that will fail to compile unless overridden.

Also introduce `kaleidoscope::hardware::avr::AVRDevice`, as a base class for our
AVR boards, that sets this up to return `Serial`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>