An intrusive boilerplate-free Focus/EEPROM storage class
This is an example of how Focus/EEPROM handling of individual
parameters could be implemented in a non-boilerplate fashion.

A new header focus_eeprom.h contains two template classes
that allow to add Focus/EEPROM handling both to members of
base classes (ReferenceFocusEEPROMWrapper)
and to define new members in derived classes
(ValueFocusEEPROMWrapper).

Using the wrapper classes is intrusive in a sense that
it requires classes (plugins) to be modified to add
Focus/EEPROM capability.

Warning: The code compiles but was not thoroughly tested.
         It is just meant as an example how things could
         be done and a basis for discussion.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>