IdleLEDs: Optional support for EEPROM-Storage & Focus
Having to recompile and re-flash firmware to set the idle timeout of the plugin
isn't a fun or desired experience. It's fine when one already figured out the
timeout they want, and have no desire to change it. For everyone else, being
able to configure it at run-time via Focus, and have it persist to EEPROM is a
much nicer experience.

This change adds an alternative version of the plugin, `PersistentIdleLEDs`, a
subclass of the original one. This one provides the focus command and
persistence. It's a child class, because that results in a smaller footprint
than a separate plugin that calls the `IdleLEDs` object.

The code borrows from - but is not wire-compatible with - Dygma's implementation
by @mattvenn.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>