New plugin: PersistentLEDMode
This implements a new plugin, `PersistentLEDMode`, whose single purpose is to
store the current LED mode to storage, whenever it changes. Since we can't hook
into led mode change events yet, we abuse the `afterEachCycle()` hook to compare
the current led mode to what we think it is, and store it if it changes.

This is obviously not very elegant, but the best we can do right now.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>