Implemented host_keymaps
host_keymaps enable defining keymaps in a natural fashion.
Instead of using Key constants like Key_A or LSHIFT(Key_A) it allows to
conveniently write "a" or "A".

The mappings between ascii and unicode characters to USB-HID keys
works by reverse engineering the host keymaps of a linux system.
The information of the provided keymap files allows for precisely
figuring out the Kaleidoscope-Key that is needed to generate a
specific utf8 character in a given keymap. For non-unicode
keycodes, the linux XKB-keysym name is mapped to a Kaleidoscope-Key.

The newly introduced host_keymap system is easily extensible and allows
users to define their own non-english keymaps, if necessary
by extending an existing keymap.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>