SAMD: Fix the descriptor sending
When sending the HID descriptor in `HID_::getDescriptor()`, we need to send it
as one package on SAMD. For this reason, we tell the USB library to pack
messages before sending anything, and then flush it at the end.

To be able to avoid ifdefs in that function, we introduce `USB_PackMessages()`,
which is a wrapper around `USBDevice.packMessages()` on SAMD, and a no-op on
AVR (where we do not need this packing).

Signed-off-by: Gergely Nagy <algernon@keyboard.io>