driver::storage::Flash: FlashAsStorage-based storage component
This is a primitive implementation of a `FlashAsStorage` (or rather,
`FlashAsEEPROM`)-based storage component. It's based on `FlashAsEEPROM`, because
I couldn't find a sane way to push the storage data variable within our template
class.

At some point, this needs to be reworked, to pull the size from Props, and not
use the EEPROM API wrappers.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>