Introduced a build type selection header
Up to now, the device header was included by several files
via the command

This commit introduces a build type selection header
kaleidoscope/device/device.h that enables to
either directly include the device header or
to first include the 'physical' device header
and then the 'virtual' device header.
This is meant for the virtual device to be able
to be defined depending on the properties of the
physical device.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>